Extract Protocol Kit Safe Factory types
